Make merkle tree commit benchmark generic over packed field
Extract the packed-field-aware commit dispatch from fri::commit_interleaved
into binius_iop_prover::merkle_tree::commit_field_buffer, and move the
binary_merkle_tree benchmark from binius-prover to binius-iop-prover, making
it generic over P: PackedField<Scalar = B128> and benching OptimalPackedB128
alongside the scalar packing.

The test_parallel_iterator test moves to merkle_tree::tests so the
to_par_scalar_* helpers can stay private.
