+++ crates/iop-prover/Cargo.toml
@@ -28,6 +28,7 @@ tracing.workspace = true
 [dev-dependencies]
 anyhow.workspace = true
 binius-math = { path = "../math", features = ["test-utils"] }
+criterion.workspace = true
 proptest.workspace = true
 rand = { workspace = true, features = ["thread_rng"] }
 sha2.workspace = true
@@ -35,3 +36,7 @@ sha2.workspace = true
 [features]
 default = ["rayon"]
 rayon = ["binius-utils/rayon"]
+
+[[bench]]
+name = "binary_merkle_tree"
+harness = false
