+++ crates/iop-prover/src/merkle_tree/mod.rs
@@ -1,6 +1,8 @@
 // Copyright 2025 Irreducible Inc.
 
+use binius_field::{Field, PackedField};
 use binius_iop::merkle_tree::{Commitment, MerkleTreeScheme};
+use binius_math::FieldSlice;
 use binius_transcript::{BufMut, TranscriptWriter};
 use binius_utils::rayon::prelude::*;
 
@@ -83,4 +85,71 @@ pub trait MerkleTreeProver<T> {
 		index: usize,
 		proof: &mut TranscriptWriter<B>,
 	);
+}
+
+/// Commits a field buffer to a Merkle tree, packing `1 << log_batch_size` scalars into each leaf.
+///
+/// The buffer's scalars are grouped into leaves of `1 << log_batch_size` elements each, in order.
+/// This dispatches on whether a leaf spans at least one full packed element (`log_batch_size >=
+/// P::LOG_WIDTH`) or fits within a single packed element, choosing the parallel scalar iterator
+/// that avoids splitting work below the packing granularity.
+///
+/// ## Preconditions
+///
+/// * The resulting leaf count (`buffer.len() / (1 << log_batch_size)`) must be a power of two.
+#[allow(clippy::type_complexity)]
+pub fn commit_field_buffer<F, P, MerkleProver>(
+	merkle_prover: &MerkleProver,
+	buffer: FieldSlice<P>,
+	log_batch_size: usize,
+) -> (
+	Commitment<<MerkleProver::Scheme as MerkleTreeScheme<F>>::Digest>,
+	MerkleProver::Committed,
+)
+where
+	F: Field,
+	P: PackedField<Scalar = F>,
+	MerkleProver: MerkleTreeProver<F>,
+{
+	if log_batch_size >= P::LOG_WIDTH {
+		let iterated_big_chunks = to_par_scalar_big_chunks(buffer.as_ref(), 1 << log_batch_size);
+		merkle_prover.commit_iterated(iterated_big_chunks)
+	} else {
+		let iterated_small_chunks =
+			to_par_scalar_small_chunks(buffer.as_ref(), 1 << log_batch_size);
+		merkle_prover.commit_iterated(iterated_small_chunks)
+	}
+}
+
+/// Creates a parallel iterator over scalars of subfield elements. Assumes chunk_size to be a power
+/// of two.
+fn to_par_scalar_big_chunks<P>(
+	packed_slice: &[P],
+	chunk_size: usize,
+) -> impl IndexedParallelIterator<Item: Iterator<Item = P::Scalar> + Send + '_>
+where
+	P: PackedField,
+{
+	packed_slice
+		.par_chunks(chunk_size / P::WIDTH)
+		.map(|chunk| PackedField::iter_slice(chunk))
+}
+
+fn to_par_scalar_small_chunks<P>(
+	packed_slice: &[P],
+	chunk_size: usize,
+) -> impl IndexedParallelIterator<Item: Iterator<Item = P::Scalar> + Send + '_>
+where
+	P: PackedField,
+{
+	(0..packed_slice.len() * P::WIDTH)
+		.into_par_iter()
+		.step_by(chunk_size)
+		.map(move |start_index| {
+			let packed_item = &packed_slice[start_index / P::WIDTH];
+			packed_item
+				.iter()
+				.skip(start_index % P::WIDTH)
+				.take(chunk_size)
+		})
 }
